+++ src/components/DesignCard.js
@@ -1,52 +1,8 @@
-import React, { useEffect, useLayoutEffect, useState} from 'react';
+import React, { useEffect } from 'react';
 
-const DesignCard = ({design}) => {
-    const [docHeight, setDocHeight] = useState(0);
+const DesignCard = ({design,docHeight}) => {
 
     var pos = Math.floor(Math.random() * 5).toString();
-
-    // animate lines
-    useEffect(() => {
-        var cardContainer = document.getElementsByClassName("cardContainer")[0].parentElement;
-        setDocHeight(cardContainer.offsetHeight);
-        console.log(cardContainer.offsetHeight);
-
-        const path = document.getElementById('line-path');
-        var pathLength = 0
-        
-        var lastScrollTop = 0;
-        var scrollUp = false;
-
-        process.nextTick(() => {
-            pathLength = path.getTotalLength()
-            path.style.strokeDasharray = pathLength + ' ' + pathLength;
-            path.style.strokeDashoffset = pathLength; 
-        });
-        
-
-        var drawLength = 0;
-        var lastDrawLength = 0;
-    
-        window.onscroll = () => {
-            var st = window.pageYOffset || document.documentElement.scrollTop;
-
-            // only apply if scrolling down
-            if(st > lastScrollTop){
-                // DO NOT RESET
-                drawLength = pathLength * ((document.documentElement.scrollTop + document.body.scrollTop) / (document.documentElement.scrollHeight - document.documentElement.clientHeight));
-                if(!scrollUp || drawLength >= lastDrawLength){
-                    console.log("p: " + pathLength.toString() + "d: " + drawLength.toString());
-                    path.style.strokeDashoffset = pathLength - drawLength;
-                }
-            }else{
-                scrollUp = true;
-                lastDrawLength = drawLength;
-            }
-            
-            // for mobile devices
-            lastScrollTop = st <= 0 ? 0 : st;
-        }     
-      }, []);
 
     if(design.type === 1){
         return(
@@ -54,10 +10,10 @@ const DesignCard = ({design}) => {
 
             
             <div className="imgContainer flex-1 relative">
-                <svg viewbox={"0 0 100 100" /* + docHeight.toString() */} className="absolute z-0" id="line-svg" height={docHeight}>
-                    <line x1="50%" y1="0%" x2="50%" y2="100%"  id="line-path" fill="none" stroke="white" stroke-width="1" />
+                <svg viewBox={"0 0 100 " + docHeight} className="absolute z-0" id="line-svg" height={docHeight}>
+                    <line x1="50%" y1="0%" x2="50%" y2="100%"  id="line-path" fill="none" stroke="white" strokeWidth="1" />
                 </svg>
-                <img class="absolute z-10" src={design.img}/>      
+                <img className="absolute z-10" src={design.img}/>      
             </div>
             <div className="textContainer flex-1"><p>{design.text_1}</p></div>
         </div>
